+++ golb/src/app/Backend/Document/note.txt
@@ -152,3 +152,6 @@ aws ecr get-login-password --region eu-west-2 \
   | docker login --username AWS --password-stdin $AWS_ACCOUNT_ID.dkr.ecr.eu-west-2.amazonaws.com
 
 docker push $AWS_ACCOUNT_ID.dkr.ecr.eu-west-2.amazonaws.com/frontend:latest
+
+
+#naming convention si important in docker
