+++ golb/src/app/Frontend/CreatePost/page.js
@@ -7,9 +7,9 @@ import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Textarea } from "@/components/ui/textarea";
 import Link from "next/link";
-import PageWrapper from "@/app/Components/PageWrapper";
-import CardContainer from "@/app/Components/CardContainer";
-import PageHeader from "@/app/Components/PageHeader";
+import PageWrapper from "@/app/Components2/PageWrapper";
+import CardContainer from "@/app/Components2/CardContainer";
+import PageHeader from "@/app/Components2/PageHeader";
 
 export default function CreatePost() {
   const [title, setTitle] = useState("");
