+++ golb/src/app/Frontend/Forms/ForgottenPassword/page.js
@@ -5,9 +5,9 @@ import { useState } from "react";
 import axios from "axios";
 import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
-import PageWrapper from "@/app/components/PageWrapper";
-import CardContainer from "@/app/components/CardContainer";
-import PageHeader from "@/app/components/PageHeader";
+import PageWrapper from "@/app/Components2/PageWrapper";
+import CardContainer from "@/app/Components2/CardContainer";
+import PageHeader from "@/app/Components2/PageHeader";
 
 export default function ForgotPasswordForm() {
   const [email, setEmail] = useState("");
