+++ golb/src/app/Frontend/Profile/[username]/page.js
@@ -4,12 +4,12 @@ import { useEffect, useState } from "react";
 import { usePathname } from "next/navigation";
 import axios from "axios";
 import Link from "next/link";
-import PostCard from "@/app/Components/PostCard";
-import FollowButton from "@/app/Components/FollowButton";
+import PostCard from "@/app/Components2/PostCard";
+import FollowButton from "@/app/Components2/FollowButton";
 import { Button } from "@/components/ui/button";
-import PageWrapper from "@/app/components/PageWrapper";
-import CardContainer from "@/app/components/CardContainer";
-import PageHeader from "@/app/components/PageHeader";
+import PageWrapper from "@/app/Components2/PageWrapper";
+import CardContainer from "@/app/Components2/CardContainer";
+import PageHeader from "@/app/Components2/PageHeader";
 
 export default function ProfilePage() {
   const pathname = usePathname();
